requirements.txt
beautifulsoup4
requests
pillow
pyyaml
cairosvg